requirements.txt
requests>=2.25.0
plyvel>=1.3.0
cryptography>=3.4.0
base58>=2.1.0
bech32>=1.2.0
bip-utils>=2.7.0
eth-keys>=0.4.0
eth-utils>=2.0.0
solders>=0.16.0
solana>=0.30.0
PyNaCl>=1.5.0
debugpy>=1.8.0
